backend/pyproject.toml
[project]
name = "layak-backend"
version = "0.1.0"
description = "Layak backend — FastAPI + ADK-Python agent pipeline for Malaysian social-assistance schemes."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "python-multipart>=0.0.9",
  "google-adk>=1.31,<1.32",
  "google-genai>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.6",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "ASYNC"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
